Moved the Request class into a separate Swift file.
